Implement the function: get_dot_definition; creating a point class
