--- enter_coordinates.h
@@ -1,7 +1,0 @@
-#ifndef ENTER_COORDINATES.H
-#define ENTER_COORDINATES.H
-
-
-int* enter_coordinates();
-
-#endif
